Modify btree index reader and writer to support 2g rows (#495)
## What changes were proposed in this pull request?

1. Previously we write row id into a buffer and use writer to write them in oneshot, which may overflow for as many as 2g rows. Currently modified to writing them directly. 
2. Index file format. The row id list size set to long.
3. Reading index fiber in memory manager. Now use a multi-read style to avoid overflowing.

## How was this patch tested?

Test on 2g rows table in one partition for multi queries. There should be a test suite for this but the least memory requirement is 8G, so this needs some discussion here.


* modify btree index reader and writer to support 2g rows

* code clean and reorganize

* code clean

* add assert and testes all done

+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/filecache/MemoryManager.scala
@@ -188,7 +188,7 @@ private[oap] object MemoryManager extends Logging {
   def cacheMemory: Long = _cacheMemory
   def cacheGuardianMemory: Long = _cacheGuardianMemory
 
-  private[filecache] def allocate(numOfBytes: Int): MemoryBlock = {
+  private[filecache] def allocate(numOfBytes: Long): MemoryBlock = {
     _memoryUsed.getAndAdd(numOfBytes)
     logDebug(s"allocate $numOfBytes memory, used: $memoryUsed")
     MemoryAllocator.UNSAFE.allocate(numOfBytes)
@@ -205,7 +205,6 @@ private[oap] object MemoryManager extends Logging {
   def putToIndexFiberCache(in: FSDataInputStream, position: Long, length: Int): IndexFiberCache = {
     val bytes = new Array[Byte](length)
     in.readFully(position, bytes)
-
     val memoryBlock = allocate(bytes.length)
     Platform.copyMemory(
       bytes,
